Move override mechanism parts to a dedicated component directory
