+++ lib/ts/recipe/emailpassword/components/themes/signInAndUp/signUpFooter.tsx
@@ -21,7 +21,7 @@ import { jsx } from "@emotion/react";
 
 import { useContext } from "react";
 import StyleContext from "../../../../../styles/styleContext";
-import { withOverride } from "../../../../withOverride";
+import { withOverride } from "../../../../../components/componentOverride/withOverride";
 
 export default withOverride(
     "EmailPasswordSignUpFooter",
