+++ lib/ts/recipe/emailpassword/components/themes/signInAndUp/signUpForm.tsx
@@ -20,7 +20,7 @@ import { jsx } from "@emotion/react";
 import { SignUpThemeProps } from "../../../types";
 
 import FormBase from "../../library/formBase";
-import { withOverride } from "../../../../withOverride";
+import { withOverride } from "../../../../../components/componentOverride/withOverride";
 
 /*
  * Component.
